@@ -8,24 +8,29 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5|~8.0.0",
-        "cloudinary-labs/cloudinary-laravel": "1.0.4",
+        "php": "^8.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "cloudinary-labs/cloudinary-laravel": "^1.0.4",
         "fideloper/proxy": "^4.4",
-        "laravel/framework": "^6.20.26",
-        "laravel/tinker": "^2.5"
+        "laravel/framework": "^9.0",
+        "laravel/tinker": "^2.5",
+        "pusher/pusher-php-server": "^5.0"
     },
     "require-dev": {
-        "facade/ignition": "^1.16.15",
+        "spatie/laravel-ignition": "^1.0",
         "fakerphp/faker": "^1.9.1",
-        "laravel/ui": "^1.0",
+        "laravel/ui": "^3.0",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^3.0",
+        "nunomaduro/collision": "^6.1",
         "phpunit/phpunit": "^8.5.8|^9.3.3"
     },
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
-        "sort-packages": true
+        "sort-packages": true,
+        "platform": {
+            "php": "8.0.21"
+        }
     },
     "extra": {
         "laravel": {
