@@ -14,7 +14,7 @@ class CreatePostRepliesTable extends Migration
     public function up()
     {
         Schema::create('post_replies', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->id();
             $table->string('body');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('post_id');
