@@ -22,7 +22,7 @@
         
     </head>
     <body>
-        <!-- Navigation-->
+        <div id='app'></div>
         <nav class="navbar navbar-expand-lg navbar-light" id="mainNav">
             <div class="container px-4 px-lg-5">
                 <a class="navbar-brand" href="/">DoggieCat</a>
